requirements.txt
openai>=1.0.0
pdfminer.six
tqdm